Regrouping duplication declarations' test files wrongly grouped
