+++ testfiles/duplicationsconflicting/left/Test.java
@@ -1,4 +1,14 @@
 public class Test {
-	int attr;
-	int otherattr;
+
+    long fibo(int n) {
+        if (n < 2) {
+            return n;
+        } else {
+            return fibo(n - 1) + fibo(n - 2);
+        }
+    }
+
+	int calc(int a,int b){
+		return a+b;
+	}
 }
