+++ testfiles/duplicationsconflicting/right/Test.java
@@ -1,4 +1,15 @@
 public class Test {
-	int otherattr;
-	int attr;
+
+	int calc(int a,int b){
+		return a*b;
+	}
+
+    long fibo(int n) {
+        if (n < 2) {
+            return n;
+        } else {
+            return fibo(n - 1) + fibo(n - 2);
+        }
+    }
+
 }
