+++ testfiles/duplicationsnoconflict/base/Test.java
@@ -1,11 +1,2 @@
 public class Test {
-
-    long fibo(int n) {
-        if (n < 2) {
-            return n;
-        } else {
-            return fibo(n - 1) + fibo(n - 2);
-        }
-    }
-
 }
