+++ testfiles/duplicationsnoconflict/left/Test.java
@@ -1,14 +1,4 @@
 public class Test {
-
-    long fibo(int n) {
-        if (n < 2) {
-            return n;
-        } else {
-            return fibo(n - 1) + fibo(n - 2);
-        }
-    }
-
-	int calc(int a,int b){
-		return a+b;
-	}
+	int attr;
+	int otherattr;
 }
